Added some more unit tests and added live updating of patient status

+++ MediQuick/MediQuick.Web/Controllers/AmbulanceController.cs
@@ -21,15 +21,27 @@ namespace MediQuick.Web.Controllers
             SetUpBaseModel(model);
 
             model.Ambulance = ambulanceService.GetAmbulanceByUserId(model.User.Id);
+            
+            if (model.Ambulance.PatientId != null)
+            {
+                model.FirstName = model.Ambulance.Patient.FirstName;
+                model.LastName = model.Ambulance.Patient.LastName;
+                model.Status = model.Ambulance.Patient.Status;
+                model.DateOfBirth = model.Ambulance.Patient.DateOfBirth;
+                model.ExtraInfo = model.Ambulance.Patient.ExtraInfo;
+                model.Sex = model.Ambulance.Patient.Sex;
+                model.SocialSecurityNumber = model.Ambulance.Patient.SocialSecurityNumber;
+            }
 
             return View(model);
         }
 
+
         [HttpPost]
-        public IActionResult AssignPatientPost(AssignPatientModel model)
+        public IActionResult AssignPatientPost(AssignPatientModel model, string buttonValue)
         {
             SetUpBaseModel(model);
-            
+
             if (model.User.UsersRoles.Select(x => x.Role.Name).Contains(RoleType.AmbulanceDriver.ToString()))
             {
                 model.Ambulance = ambulanceService.GetAmbulanceByUserId(model.User.Id);
@@ -39,19 +51,69 @@ namespace MediQuick.Web.Controllers
                 model.Ambulance = ambulanceService.GetAmbulanceById(model.AmbulanceId);
             }
 
-            try
+            if (buttonValue == "AssignPatientPost")
             {
-                if (model.Ambulance.PatientId != null)
-                {
-                    ambulanceService.RemoveAmbulancePatient(model.Ambulance.Id);
+                
 
-                    AssignPatientModel baseModel = new AssignPatientModel();
-                    baseModel.Messages.Add(new Message("Patient was unassigned", MessageType.Success));
+                try
+                {
+                    if (model.Ambulance.PatientId != null)
+                    {
+                        ambulanceService.RemoveAmbulancePatient(model.Ambulance.Id);
+
+
+                        model.FirstName = "";
+                        model.LastName = "";
+                        model.SocialSecurityNumber = "";
+                        model.Sex = "";
+                        model.Status = "";
+                        model.DateOfBirth = DateTime.Now;
+                        model.ExtraInfo = "";
+
+                        ModelState.Clear();
+
+                        SetUpBaseModel(model);
+
+                        model.Messages.Add(new Message("Patient was unassigned", MessageType.Success));
+                        return View("AssignPatient", model);
+                    }
+                    else
+                    {
+                        ambulanceService.AssignPatientToAmbulance(
+                            model.Ambulance.Id,
+                            model.FirstName,
+                            model.LastName,
+                            model.SocialSecurityNumber,
+                            model.Sex,
+                            model.Status,
+                            model.DateOfBirth,
+                            model.ExtraInfo);
+
+                        
+                        
+                        model.Messages.Add(new Message("Patient was assigned", MessageType.Success));
+                        return View("AssignPatient", model);
+                    }
+                }
+                catch (Exception)
+                {
+                    model = new AssignPatientModel();
+
+                    model.Messages.Add(new Message("There was an error!", MessageType.Error));
                     return View("AssignPatient", model);
                 }
-                else
+            }
+            else if(buttonValue == "UpdatePatientPost")
+            {
+                if (model.Ambulance.PatientId == null)
                 {
-                    ambulanceService.AssignPatientToAmbulance(
+                    model = new AssignPatientModel();
+
+                    model.Messages.Add(new Message("Patient was not found", MessageType.Error));
+                    return View("AssignPatient", model);
+                }
+
+                ambulanceService.UpdatePatientOfAmbulance(
                         model.Ambulance.Id,
                         model.FirstName,
                         model.LastName,
@@ -61,20 +123,12 @@ namespace MediQuick.Web.Controllers
                         model.DateOfBirth,
                         model.ExtraInfo);
 
-                    AssignPatientModel baseModel = new AssignPatientModel();
-                    baseModel.Messages.Add(new Message("Patient was assigned", MessageType.Success));
-                    return View("AssignPatient", model);
-                }
-            }
-            catch (Exception)
-            {
-                model = new AssignPatientModel();
 
-                model.Messages.Add(new Message("There was an error!", MessageType.Error));
-                return View(model);
+                model.Messages.Add(new Message("Patient info was updated", MessageType.Success));
+                return View("AssignPatient", model);
             }
 
-            return View(model);
+            return View("AssignPatient", model);
         }
 
         public IActionResult ViewAmbulance(ViewAmbulanceModel model)
